chore: move lowest level of commands one level up
