+++ cmd/jsonnet/format.go
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package format
+package jsonnet
 
 import (
 	"fmt"
@@ -60,10 +60,6 @@ Use -w or --write to write output back to files instead of stdout.
 			}
 		}
 	},
-}
-
-func RegisterCommandRecursive(parent *cobra.Command) {
-	parent.AddCommand(jsonnetFormatCmd)
 }
 
 func init() {
