+++ cmd/jsonnet/jsonnet.go
@@ -2,9 +2,6 @@ package jsonnet
 
 import (
 	"github.com/spf13/cobra"
-
-	"github.com/ory/kratos/cmd/jsonnet/format"
-	"github.com/ory/kratos/cmd/jsonnet/lint"
 )
 
 // jsonnetCmd represents the jsonnet command
@@ -16,6 +13,6 @@ var jsonnetCmd = &cobra.Command{
 func RegisterCommandRecursive(parent *cobra.Command) {
 	parent.AddCommand(jsonnetCmd)
 
-	format.RegisterCommandRecursive(jsonnetCmd)
-	lint.RegisterCommandRecursive(jsonnetCmd)
+	jsonnetCmd.AddCommand(jsonnetFormatCmd)
+	jsonnetCmd.AddCommand(jsonnetLintCmd)
 }
