+++ cmd/migrate/migrate.go
@@ -2,8 +2,6 @@ package migrate
 
 import (
 	"github.com/spf13/cobra"
-
-	"github.com/ory/kratos/cmd/migrate/sql"
 )
 
 // migrateCmd represents the migrate command
@@ -15,5 +13,5 @@ var migrateCmd = &cobra.Command{
 func RegisterCommandRecursive(parent *cobra.Command) {
 	parent.AddCommand(migrateCmd)
 
-	sql.RegisterCommandRecursive(migrateCmd)
+	migrateCmd.AddCommand(migrateSqlCmd)
 }
